+++ lib/get-package-install-command.d.ts
@@ -1,0 +1,6 @@
+import type { ReadmeConfig } from './index.d.ts';
+
+/**
+ * Get the appropriate package install command.
+ */
+export default function getPackageInstallCommand(config?: ReadmeConfig): string;