+++ lib/get-package-manager.d.ts
@@ -1,0 +1,9 @@
+import type { PackageJson } from 'type-fest';
+
+export type PackageManager = 'npm' | 'pnpm' | 'yarn';
+
+/**
+ * Get the name of the preferred package manager from the `package.json` engines
+ * property. Defaults to "npm".
+ */
+export default function getPackageManager(engines?: PackageJson.engines): PackageManager;