+++ lib/get-package-test-command.d.ts
@@ -1,0 +1,6 @@
+import type { ReadmeConfig } from './index.d.ts';
+
+/**
+ * Get the appropriate package test command.
+ */
+export default function getPackageTestCommand(config?: ReadmeConfig): string;